requirements.txt
absl-py==0.1.10
appnope==0.1.0
astroid==1.6.1
bleach==1.5.0
cycler==0.10.0
decorator==4.2.1
dj-database-url==0.5.0
Django==2.0.3
django-heroku==0.3.1
djangorestframework==3.7.7
entrypoints==0.2.3
geopy==1.11.0
gsl==0.0.1
gunicorn==19.7.1
html5lib==0.9999999
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
isort==4.3.4
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
mlpy==3.5.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.0
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pew==1.1.2
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pipenv==11.1.4
prompt-toolkit==1.0.15
protobuf==3.5.1
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.8.2
pymongo==3.6.1
pyparsing==2.2.0
PyQt5==5.10
python-dateutil==2.6.1
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
sip==4.19.7
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.8.1
testpath==0.3.1
tornado==5.0
traitlets==4.3.2
virtualenv==15.1.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==3.3.1
widgetsnbextension==3.1.4
wrapt==1.10.11
